fix(render): don't truncate flags to uint8_t in resolve()
RenderDescriptor.flags was widened to uint16_t to carry RF_Sanctum
(bit 8), but the local `flags` variable in resolve() stayed uint8_t —
silently dropping bits 8+. Sanctum walls/floors never lit up because
the RF_Sanctum check always failed. Widened the local to uint16_t.

resolve_fixture's `uint8_t flags` parameter is unchanged: it only ever
reads RF_Remembered and RF_Open, both in the low byte.

+++ src/terminal_theme.cpp
@@ -1305,7 +1305,7 @@ ResolvedVisual resolve(const RenderDescriptor& desc) {
 
     auto tile = static_cast<Tile>(desc.type_id);
     uint8_t seed = desc.seed;
-    uint8_t flags = desc.flags;
+    uint16_t flags = desc.flags;
     Biome biome = desc.biome;
     bool remembered = (flags & RF_Remembered) != 0;
